% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker_winner.R
\name{poker_winner}
\alias{poker_winner}
\title{Determine winner of a poker game}
\usage{
poker_winner()
}
\value{
A number indicating player 1 or player 2
}
\description{
Determine winner of a poker game
}
\examples{
poker_winner()
# user will be prompted to enter cards as character vector for each player
}
