% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poker_winner.R
\name{poker_winner}
\alias{poker_winner}
\title{Determine winner of a poker game}
\usage{
poker_winner(hands)
}
\arguments{
\item{hands}{list of cards that both player have in hand}
}
\value{
A number indicating winner - player 1 or player 2 or tie(0)
}
\description{
Determine winner of a poker game
}
\examples{
hands <- list(c("2h","2h","2h","2h","3h"),c("4h","4h","4h","4h","5h"))
poker_winner(hands)
}
